Extract delete lines code to custom hook
