+++ packages/invoices/src/OutboundShipment/DetailView/Toolbar.tsx
@@ -8,8 +8,6 @@ import {
   DropdownMenu,
   useTranslation,
   useBufferState,
-  useTableStore,
-  useNotification,
   DropdownMenuItem,
   DeleteIcon,
 } from '@openmsupply-client/common';
@@ -17,61 +15,17 @@ import { NameSearchInput } from '@openmsupply-client/system';
 import {
   useOutboundFields,
   useIsOutboundDisabled,
-  useDeleteInboundLine,
-  useOutboundRows,
+  useDeleteSelectedLines,
 } from '../api';
-import { InvoiceItem, InvoiceLine } from '../../types';
 
 export const Toolbar: FC = () => {
-  const { success, info } = useNotification();
-  const { items, lines } = useOutboundRows();
-  const { mutate } = useDeleteInboundLine();
+  const { onDelete } = useDeleteSelectedLines();
   const { otherParty, theirReference, update } = useOutboundFields([
     'otherParty',
     'theirReference',
   ]);
   const [theirReferenceBuffer, setTheirReferenceBuffer] =
     useBufferState(theirReference);
-
-  const { selectedRows } = useTableStore(state => {
-    const { isGrouped } = state;
-
-    if (isGrouped) {
-      return {
-        selectedRows: (
-          Object.keys(state.rowState)
-            .filter(id => state.rowState[id]?.isSelected)
-            .map(selectedId => items?.find(({ id }) => selectedId === id))
-            .filter(Boolean) as InvoiceItem[]
-        )
-          .map(({ lines }) => lines)
-          .flat()
-          .map(({ id }) => id),
-      };
-    } else {
-      return {
-        selectedRows: (
-          Object.keys(state.rowState)
-            .filter(id => state.rowState[id]?.isSelected)
-            .map(selectedId => lines.find(({ id }) => selectedId === id))
-            .filter(Boolean) as InvoiceLine[]
-        ).map(({ id }) => id),
-      };
-    }
-  });
-
-  const deleteAction = async () => {
-    if (selectedRows && selectedRows?.length > 0) {
-      const number = selectedRows?.length;
-      const onSuccess = success(t('message.deleted-lines', { number }));
-      mutate(selectedRows, {
-        onSuccess,
-      });
-    } else {
-      const infoSnack = info(t('label.select-rows-to-delete-them'));
-      infoSnack();
-    }
-  };
 
   const isDisabled = useIsOutboundDisabled();
   const t = useTranslation('distribution');
@@ -120,7 +74,7 @@ export const Toolbar: FC = () => {
           </Box>
         </Grid>
         <DropdownMenu disabled={isDisabled} label={t('label.select')}>
-          <DropdownMenuItem IconComponent={DeleteIcon} onClick={deleteAction}>
+          <DropdownMenuItem IconComponent={DeleteIcon} onClick={onDelete}>
             {t('button.delete-lines')}
           </DropdownMenuItem>
         </DropdownMenu>
